Add USB HID implementation 7

+++ firmware/src/util/common.h
@@ -6,11 +6,11 @@
 
 namespace threeboard {
 namespace util {
-__always_inline constexpr uint8_t min(uint8_t a, uint8_t b) {
+__always_inline constexpr uint16_t min(uint16_t a, uint16_t b) {
   return a < b ? a : b;
 }
 
-__always_inline constexpr uint8_t max(uint8_t a, uint8_t b) {
+__always_inline constexpr uint16_t max(uint16_t a, uint16_t b) {
   return a < b ? b : a;
 }
